requirements.txt
Django
psycopg2-binary
Pillow
django-crispy-forms
crispy-bootstrap5
python-dotenv
django-filter
